remove syncing last bars
