+++ src/domain/tradingview/useTVDatafeed.ts
@@ -23,9 +23,7 @@ export default function useTVDatafeed({ dataProvider }: Props) {
   const { chainId } = useChainId();
   const intervalRef = useRef<ReturnType<typeof setInterval> | undefined>();
   const resetCacheRef = useRef<() => void | undefined>();
-  const feedData = useRef<boolean>(true);
   const activeTicker = useRef<string | undefined>();
-  const lastTimestamp = useRef<number>(0);
   const tvDataProvider = useRef<TVDataProvider>();
 
   useEffect(() => {
@@ -33,22 +31,6 @@ export default function useTVDatafeed({ dataProvider }: Props) {
       tvDataProvider.current = dataProvider;
     }
   }, [dataProvider]);
-
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === "hidden") {
-        feedData.current = false;
-      } else {
-        feedData.current = true;
-      }
-    };
-
-    document.addEventListener("visibilitychange", handleVisibilityChange);
-
-    return () => {
-      document.removeEventListener("visibilitychange", handleVisibilityChange);
-    };
-  }, []);
 
   return useMemo(() => {
     return {
@@ -111,7 +93,6 @@ export default function useTVDatafeed({ dataProvider }: Props) {
             onErrorCallback("Unable to load historical data!");
           }
         },
-
         async subscribeBars(
           symbolInfo: SymbolInfo,
           resolution: ResolutionString,
@@ -126,25 +107,12 @@ export default function useTVDatafeed({ dataProvider }: Props) {
           intervalRef.current && clearInterval(intervalRef.current);
           resetCacheRef.current = onResetCacheNeededCallback;
           if (!isStable) {
-            intervalRef.current = setInterval(async function () {
-              if (!feedData.current) return;
-              tvDataProvider.current
-                ?.syncPastBars(chainId, ticker, resolution, lastTimestamp.current)
-                .then((bars) => {
-                  if (bars?.length > 0 && ticker === activeTicker.current) {
-                    bars.forEach((bar) => {
-                      onRealtimeCallback(bar);
-                    });
-                  }
-                })
-                .then(() => {
-                  tvDataProvider.current?.getLiveBar(chainId, ticker, resolution).then((bar) => {
-                    if (bar && ticker === activeTicker.current) {
-                      lastTimestamp.current = bar.time;
-                      onRealtimeCallback(formatTimeInBarToMs(bar));
-                    }
-                  });
-                });
+            intervalRef.current = setInterval(function () {
+              tvDataProvider.current?.getLiveBar(chainId, ticker, resolution).then((bar) => {
+                if (bar && ticker === activeTicker.current) {
+                  onRealtimeCallback(formatTimeInBarToMs(bar));
+                }
+              });
             }, 500);
           }
         },
